# src/swagger/swagger-latest.yaml
swagger: '2.0'
info:
  contact: {}
paths:
  /api/v2/lists:
    post:
      description: create list
      consumes:
        - application/json
      tags:
        - lists/v2
      summary: Create
      parameters:
        - description: List body
          name: input
          in: body
          required: true
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ListCreateRequest'
        - type: string
          description: JWT-format token
          name: X-User-Token
          in: header
          required: true
      responses:
        '201':
          description: id
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.IDResponse'
        '400':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        '401':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        '500':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
  /api/v2/lists-root:
    get:
      description: get user root list
      consumes:
        - application/json
      tags:
        - lists/v2
      summary: Get user root list
      parameters:
        - type: string
          description: JWT-format token
          name: X-User-Token
          in: header
          required: true
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ListResponse'
        '400':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        '401':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        '404':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        '500':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
  /api/v2/lists/{id}:
    get:
      description: get list
      consumes:
        - application/json
      tags:
        - lists/v2
      summary: Get
      parameters:
        - type: string
          description: JWT-format token
          name: X-User-Token
          in: header
          required: true
        - type: integer
          description: ListId
          name: id
          in: path
          required: true
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ListResponse'
        '400':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        '401':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        '404':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        '500':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
    put:
      description: update list
      consumes:
        - application/json
      tags:
        - lists/v2
      summary: Update
      parameters:
        - description: List body
          name: input
          in: body
          required: true
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ListUpdateRequest'
        - type: integer
          description: ListId
          name: id
          in: path
          required: true
        - type: string
          description: JWT-format token
          name: X-User-Token
          in: header
          required: true
      responses:
        '200':
          description: OK
        '400':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        '401':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        '404':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        '500':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
    delete:
      description: delete list
      consumes:
        - application/json
      tags:
        - lists/v2
      summary: Delete
      parameters:
        - type: string
          description: JWT-format token
          name: X-User-Token
          in: header
          required: true
        - type: integer
          description: ListId
          name: id
          in: path
          required: true
      responses:
        '200':
          description: OK
        '400':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        '401':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        '404':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        '500':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
...
  /api/v2/sublists/{id}:
    get:
      description: get sublists of the list
      consumes:
        - application/json
      tags:
        - lists/v2
      summary: Get sublists
      parameters:
        - type: string
          description: JWT-format token
          name: X-User-Token
          in: header
          required: true
        - type: integer
          description: ListId
          name: id
          in: path
          required: true
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ListResponse'
        '400':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        '401':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        '500':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
definitions:
  git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse:
    type: object
    properties:
      error:
        type: string
        example: error description
  git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.IDResponse:
    type: object
    properties:
      id:
        type: integer
  git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ListCreateRequest:
    type: object
    properties:
      name:
        type: string
      parentId:
        type: integer
...
  git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ListResponse:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      parentId:
        type: integer
      userId:
        type: integer
  git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ListUpdateRequest:
    type: object
    properties:
      name:
        type: string
...